server/requirements/base.txt
python-decouple==3.6
Django==4.1
django-phonenumber-field==5.2.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
phonenumbers==8.12.53
PyJWT==2.4.0
Pillow==9.2.0
django-cors-headers==3.13.0
redis==4.3.4
drf-yasg==1.21.3
django-jazzmin==2.5.0
django-filter==22.1
redis==4.3.4
beautifulsoup4==4.11.1
requests==2.28.1
celery==5.2.7
pusher==3.2.0
psycopg2==2.9.3
gunicorn==20.1.0
